v1.2
added async leaguetrophylistcommand, structure for api calls and cache for autocomplete

+++ src/main/java/commands/kickpoints/kpremove.java
@@ -1,11 +1,12 @@
 package commands.kickpoints;
 
+import datautil.DBUtil;
+import datawrapper.Player;
+import datawrapper.Kickpoint;
+import datawrapper.User;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
-import sql.DBUtil;
-import sql.Kickpoint;
-import sql.User;
 import util.MessageUtil;
 
 public class kpremove extends ListenerAdapter {
@@ -16,16 +17,6 @@ public class kpremove extends ListenerAdapter {
 			return;
 		event.deferReply().queue();
 		String title = "Kickpunkte";
-
-		User userexecuted = new User(event.getUser().getId());
-		if (!(userexecuted.getPermissions() == User.PermissionType.ADMIN
-				|| userexecuted.getPermissions() == User.PermissionType.LEADER
-				|| userexecuted.getPermissions() == User.PermissionType.COLEADER)) {
-			event.getHook().editOriginalEmbeds(MessageUtil.buildEmbed(title,
-					"Du musst mindestens Vize-Anführer eines Clans sein, um diesen Befehl ausführen zu können.",
-					MessageUtil.EmbedType.ERROR)).queue();
-			return;
-		}
 
 		OptionMapping idOption = event.getOption("id");
 
@@ -40,15 +31,37 @@ public class kpremove extends ListenerAdapter {
 
 		Kickpoint kp = new Kickpoint(id);
 
+		String clantag = kp.getPlayer().getClanDB().getTag();
+
+		if(clantag.equals("warteliste")) {
+			event.getHook().editOriginalEmbeds(
+					MessageUtil.buildEmbed(title, "Diesen Befehl kannst du nicht auf die Warteliste ausführen.", MessageUtil.EmbedType.ERROR))
+					.queue();
+			return;
+		}
+		
+		User userexecuted = new User(event.getUser().getId());
+		if (!(userexecuted.getClanRoles().get(clantag) == Player.RoleType.ADMIN
+				|| userexecuted.getClanRoles().get(clantag) == Player.RoleType.LEADER
+				|| userexecuted.getClanRoles().get(clantag) == Player.RoleType.COLEADER)) {
+			event.getHook()
+			.editOriginalEmbeds(MessageUtil.buildEmbed(title,
+					"Du musst mindestens Vize-Anführer des Clans sein, um diesen Befehl ausführen zu können.",
+					MessageUtil.EmbedType.ERROR)).queue();
+			return;
+		}
+
 		String desc = "";
 
 		if (kp.getDescription() != null) {
 			DBUtil.executeUpdate("DELETE FROM kickpoints WHERE id = ?", id);
 			desc += "Der Kickpunkt mit der ID " + id + " wurde gelöscht.";
-			event.getHook().editOriginalEmbeds(MessageUtil.buildEmbed(title, desc, MessageUtil.EmbedType.SUCCESS)).queue();
+			event.getHook().editOriginalEmbeds(MessageUtil.buildEmbed(title, desc, MessageUtil.EmbedType.SUCCESS))
+					.queue();
 		} else {
 			desc += "Ein Kickpunkt mit dieser ID existiert nicht.";
-			event.getHook().editOriginalEmbeds(MessageUtil.buildEmbed(title, desc, MessageUtil.EmbedType.ERROR)).queue();
+			event.getHook().editOriginalEmbeds(MessageUtil.buildEmbed(title, desc, MessageUtil.EmbedType.ERROR))
+					.queue();
 		}
 
 	}
